# backend/src/main/java/com/yuuka/backend/auth/api/dto/LoginRequest.java
package com.yuuka.backend.auth.api.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;
import jakarta.validation.constraints.Size;

public record LoginRequest(
    @Email @NotBlank @Size(max = 320) String email,
    @NotBlank @Size(min = 8, max = 128) String password,
    @Pattern(regexp = "^$|\\d{6}$", message = "must be a 6 digit code") String totpCode) {}