# backend/src/main/java/com/yuuka/backend/auth/api/dto/LogoutRequest.java
package com.yuuka.backend.auth.api.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record LogoutRequest(@NotBlank @Size(max = 256) String refreshToken) {}