# backend/src/main/java/com/yuuka/backend/auth/api/dto/RefreshRequest.java
package com.yuuka.backend.auth.api.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record RefreshRequest(@NotBlank @Size(max = 256) String refreshToken) {}