# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Yuuka API",
    "description" : "Private paycheck-first budgeting API.",
    "version" : "v1"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
...
    "/api/v1/entries/{entryId}/bucket-transactions" : {
      "get" : {
        "tags" : [ "bucket-transaction-controller" ],
        "operationId" : "list_1",
        "parameters" : [ {
          "name" : "entryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseBucketTransactionResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "bucket-transaction-controller" ],
        "operationId" : "create_2",
        "parameters" : [ {
          "name" : "entryId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateBucketTransactionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BucketTransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/register" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "summary" : "Register an account when explicitly enabled",
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "summary" : "Rotate a refresh token and issue a new session",
        "operationId" : "refresh",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "summary" : "Revoke a refresh token",
        "operationId" : "logout",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LogoutRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/v1/auth/login" : {
      "post" : {
        "tags" : [ "auth-controller" ],
        "summary" : "Authenticate the configured owner",
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/bucket-transactions/{transactionId}" : {
      "delete" : {
        "tags" : [ "bucket-transaction-controller" ],
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "transactionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "version",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      },
      "patch" : {
        "tags" : [ "bucket-transaction-controller" ],
        "operationId" : "update_2",
        "parameters" : [ {
          "name" : "transactionId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateBucketTransactionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BucketTransactionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "CreateBucketTransactionRequest" : {
        "type" : "object",
        "properties" : {
          "amountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "description" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "effectiveDate" : {
            "type" : "string",
            "format" : "date"
          }
        },
        "required" : [ "amountMinor", "effectiveDate" ]
      },
      "BucketTransactionResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "entryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "amountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "description" : {
            "type" : "string"
          },
          "effectiveDate" : {
            "type" : "string",
            "format" : "date"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "RegisterRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "maxLength" : 320,
            "minLength" : 0
          },
          "password" : {
            "type" : "string",
            "maxLength" : 128,
            "minLength" : 12,
            "pattern" : "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d).+$"
          },
          "displayName" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          }
        },
        "required" : [ "email", "password" ]
      },
      "AuthResponse" : {
        "type" : "object",
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "tokenType" : {
            "type" : "string"
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "refreshToken" : {
            "type" : "string"
          },
          "refreshExpiresAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "RefreshRequest" : {
        "type" : "object",
        "properties" : {
          "refreshToken" : {
            "type" : "string",
            "maxLength" : 256,
            "minLength" : 0
          }
        },
        "required" : [ "refreshToken" ]
      },
      "LogoutRequest" : {
        "type" : "object",
        "properties" : {
          "refreshToken" : {
            "type" : "string",
            "maxLength" : 256,
            "minLength" : 0
          }
        },
        "required" : [ "refreshToken" ]
      },
      "LoginRequest" : {
        "type" : "object",
        "properties" : {
          "email" : {
            "type" : "string",
            "maxLength" : 320,
            "minLength" : 0
          },
          "password" : {
            "type" : "string",
            "maxLength" : 128,
            "minLength" : 8
          },
          "totpCode" : {
            "type" : "string",
            "pattern" : "^$|\\d{6}$"
          }
        },
        "required" : [ "email", "password" ]
      },
...
      "UpdateBucketTransactionRequest" : {
        "type" : "object",
        "properties" : {
          "amountMinor" : {
            "type" : "integer",
            "format" : "int64"
          },
          "description" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "effectiveDate" : {
            "type" : "string",
            "format" : "date"
          },
          "version" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "amountMinor", "effectiveDate", "version" ]
      },
...
      "PageResponseBucketTransactionResponse" : {
        "type" : "object",
        "properties" : {
          "items" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BucketTransactionResponse"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalItems" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "hasNext" : {
            "type" : "boolean"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}